Parse LaTeX code

The parseLatex function parses LaTeX source, producing a structured object.

This package exports an enhanced version of the tools::parseLatex function, which handles more LaTeX syntax.

Usage

parseLatex(
  text,
  filename = deparse1(substitute(text)),
  verbose = FALSE,
  verbatim = c("verbatim", "verbatim*", "Sinput", "Soutput"),
  verb = "\\Sexpr"
)

# S3 method for class 'LaTeX2item'
print(x, ...)

# S3 method for class 'LaTeX2'
print(x, tags = FALSE, ...)

Arguments

text

A character vector containing LaTeX source code.

filename

A filename to use in syntax error messages.

verbose

If TRUE, print debug error messages.

verbatim

A character vector containing the names of LaTeX environments holding verbatim text.

verb

A character vector containing LaTeX macros that should be assumed to hold verbatim text.

x

Object to work on.

...

Extra parameters to pass to `deparseLatex`

tags

Whether to display LaTeX2 tags

Value

Parsed Latex in list with class `"LaTeX2"`.

See also

Author

Maintainer: Duncan Murdoch murdoch.duncan@gmail.com

Examples

parsed <- parseLatex(r"(fran\c{c}ais)")
parsed
#> fran\c{c}ais 
getParseData(parsed)
#> NULL