Miscellaneous low-level finders

Miscellaneous low-level finders

Usage

find_whitespace(items)

find_env(items, envtypes)

find_macro(items, macros)

find_catcode(items, codes)

find_tags(items, tags)

find_char(items, char)

Arguments

items

A list of latex items.

envtypes

Which types of environment to look for.

macros

Which types of macros to look for.

codes

Which codes to look for.

tags

Which tags to look for.

char

Which character to look for.

Value

find_whitespace() returns the indices of whitespace in items.

find_env() returns the indices within items of environments in envtypes.

find_macro() returns the index within items of instances in macros.

find_catcode() returns the index within items. of specials matching code.

find_tags() returns the index within items. of items with tags matching tags.

find_char() returns the index within items of characters matching char. Only characters marked as SPECIAL by the parser will be found.